requirements.txt

numpy>=1.21.0
networkx>=2.8.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.9.0
numba>=0.56.0
joblib>=1.2.0
tqdm>=4.64.0
sigfig>=1.3.0
rdkit>=2022.9.0
matplotlib>=3.6.0
pillow>=9.0.0

pytest>=7.2.0
